refactor: move messages into `message` module
