+++ amico-sdk/src/ai/mod.rs
@@ -1,7 +1,7 @@
-pub mod chat;
 pub mod completion;
 pub mod embedding;
 pub mod errors;
+pub mod message;
 pub mod provider;
 pub mod service;
 pub mod tool;
